requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Authentication & Security
djangorestframework-simplejwt==5.3.1
cryptography==42.0.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# AI & ML
openai==1.12.0
langchain==0.1.7
langchain-openai==0.0.5
chromadb==0.4.22
sentence-transformers==2.3.1
tiktoken==0.5.2

# Data Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.4
numpy==1.26.3

# Storage & Files
boto3==1.34.34
django-storages==1.14.2
Pillow==10.2.0

# Task Queue
celery==5.3.6
redis==5.0.1

# API & Web
requests==2.31.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3

# Development
black==24.1.1
flake8==7.0.0
pytest==8.0.0
pytest-django==4.7.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
